(SP 1) [Shop UI] Add page metadata across shop routes
